Some small changes in the Player classes.
I will try to explain the changes as clear as possible, but forgive me
for the fact my explaining is horrible :c

Basically the Player classes store an array called piecesData, which is
an Integer array storing the ranks of all pieces which you can place (0
- 11) and how many of each piece you can place (the amount is hardcoded
so far).

Furthermore, it stores the locations of the Pieces in an array of Pieces
called pieces. These arrays are stored in the Player class because the
Board needs to know what piece to place and where to place that piece.
The Pieces are created by the Piece class and basically created by the
Player, the Board just places them on the stored location.